Fix exception thrown login vs plugin
