+++ Engine/RealTime/LiveTradingRealTimeHandler.cs
@@ -24,6 +24,9 @@ using QuantConnect.Logging;
 using QuantConnect.Packets;
 using QuantConnect.Scheduling;
 using QuantConnect.Util;
+using QuantConnect.Securities;
+using System.Collections.Generic;
+using QuantConnect.Configuration;
 
 namespace QuantConnect.Lean.Engine.RealTime
 {
@@ -42,6 +45,8 @@ namespace QuantConnect.Lean.Engine.RealTime
         private IAlgorithm _algorithm;
         private IResultHandler _resultHandler;
 
+        private static MarketHoursDatabase _marketHoursDatabase;
+
         /// <summary>
         /// Boolean flag indicating thread state.
         /// </summary>
@@ -57,6 +62,7 @@ namespace QuantConnect.Lean.Engine.RealTime
             _algorithm = algorithm;
             _resultHandler = resultHandler;
             _cancellationTokenSource = new CancellationTokenSource();
+            _marketHoursDatabase = MarketHoursDatabase.FromDataFolder();
 
             var todayInAlgorithmTimeZone = DateTime.UtcNow.ConvertFromUtc(_algorithm.TimeZone).Date;
 
@@ -155,7 +161,7 @@ namespace QuantConnect.Lean.Engine.RealTime
             {
                 var security = kvp.Value;
 
-                var marketHours = _api.MarketToday(date, security.Symbol);
+                var marketHours = MarketToday(date, security.Symbol);
                 security.Exchange.SetMarketHours(marketHours, date.DayOfWeek);
                 var localMarketHours = security.Exchange.Hours.MarketHours[date.DayOfWeek];
                 Log.Trace($"LiveTradingRealTimeHandler.RefreshMarketHoursToday({security.Type}): Market hours set: Symbol: {security.Symbol} {localMarketHours} ({security.Exchange.Hours.TimeZone})");
@@ -211,6 +217,24 @@ namespace QuantConnect.Lean.Engine.RealTime
         public void Exit()
         {
             _cancellationTokenSource.Cancel();
+        }
+
+        /// <summary>
+        /// Get the calendar open hours for the date.
+        /// </summary>
+        private IEnumerable<MarketHoursSegment> MarketToday(DateTime time, Symbol symbol)
+        {
+            if (Config.GetBool("force-exchange-always-open"))
+            {
+                yield return MarketHoursSegment.OpenAllDay();
+                yield break;
+            }
+
+            var hours = _marketHoursDatabase.GetExchangeHours(symbol.ID.Market, symbol, symbol.ID.SecurityType);
+            foreach (var segment in hours.MarketHours[time.DayOfWeek].Segments)
+            {
+                yield return segment;
+            }
         }
     }
 }
